Revert "Revert "Some fix""
This reverts commit ac4ad5351cecdc43882767fe2fd3b16f854acea3.

+++ FusionSamples/Submarines/SubmarinesGameLibrary/GameActions/Accident.cs
@@ -9,18 +9,24 @@ namespace SubmarinesWars.SubmarinesGameLibrary.GameActions
 {
     class Accident : Action
     {
-        public Accident(Submarine sub, ActionsQueue queue)
+        Submarine sub1;
+
+        public Accident(Submarine sub, Submarine sub1, ActionsQueue queue)
         {
             Entity = sub;
+            this.sub1 = sub1;
             ActionsQueue = queue;
         }
 
         public override bool execute(Fusion.GameTime gameTime)
         {
-            ((Submarine)Entity).damage();
-            ((Team)Entity.Parent).AI.NotifyAccident((Submarine)Entity);
-            ActionsQueue.Field.addNewNoise(Entity.Cell, Config.NOISE_ACCIDENT);
-            if (Entity.ToRemove) ActionsQueue.deleteAllFor(Entity); 
+            if (Entity.Cell == sub1.Cell)
+            {
+                ((Submarine)Entity).damage();
+                ((Team)Entity.Parent).AI.NotifyAccident((Submarine)Entity);
+                ActionsQueue.Field.addNewNoise(Entity.Cell, Config.NOISE_ACCIDENT);
+                if (Entity.ToRemove) ActionsQueue.deleteAllFor(Entity);
+            }
             return true;
         }
     }
